+++ grype/version/apk_version_test.go
@@ -8,7 +8,60 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestApkVersionCompare(t *testing.T) {
+func TestVersionApk_Constraint(t *testing.T) {
+	tests := []testCase{
+		{version: "2.3.1", constraint: "", satisfied: true},
+		// compound conditions
+		{version: "2.3.1", constraint: "> 1.0.0, < 2.0.0", satisfied: false},
+		{version: "1.3.1", constraint: "> 1.0.0, < 2.0.0", satisfied: true},
+		{version: "2.0.0", constraint: "> 1.0.0, <= 2.0.0", satisfied: true},
+		{version: "2.0.0", constraint: "> 1.0.0, < 2.0.0", satisfied: false},
+		{version: "1.0.0", constraint: ">= 1.0.0, < 2.0.0", satisfied: true},
+		{version: "1.0.0", constraint: "> 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.9.0", constraint: "> 1.0.0, < 2.0.0", satisfied: false},
+		{version: "1.5.0", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: true},
+		{version: "0.2.0", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: true},
+		{version: "0.0.1", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.6.0", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "2.5.0", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		// fixed-in scenarios
+		{version: "2.3.1", constraint: "< 2.0.0", satisfied: false},
+		{version: "2.3.1", constraint: "< 2.0", satisfied: false},
+		{version: "2.3.1", constraint: "< 2", satisfied: false},
+		{version: "2.3.1", constraint: "< 2.3", satisfied: false},
+		{version: "2.3.1", constraint: "< 2.3.1", satisfied: false},
+		{version: "2.3.1", constraint: "< 2.3.2", satisfied: true},
+		{version: "2.3.1", constraint: "< 2.4", satisfied: true},
+		{version: "2.3.1", constraint: "< 3", satisfied: true},
+		{version: "2.3.1", constraint: "< 3.0", satisfied: true},
+		{version: "2.3.1", constraint: "< 3.0.0", satisfied: true},
+		// alpine specific scenarios
+		// https://wiki.alpinelinux.org/wiki/APKBUILD_Reference#pkgver
+		{version: "1.5.1-r1", constraint: "< 1.5.1", satisfied: false},
+		{version: "1.5.1-r1", constraint: "> 1.5.1", satisfied: true},
+		{version: "9.3.2-r4", constraint: "< 9.3.4-r2", satisfied: true},
+		{version: "9.3.4-r2", constraint: "> 9.3.4", satisfied: true},
+		{version: "4.2.52_p2-r1", constraint: "< 4.2.52_p4-r2", satisfied: true},
+		{version: "4.2.52_p2-r1", constraint: "> 4.2.52_p4-r2", satisfied: false},
+		{version: "0.1.0_alpha", constraint: "< 0.1.3_alpha", satisfied: true},
+		{version: "0.1.0_alpha2", constraint: "> 0.1.0_alpha", satisfied: true},
+		{version: "1.1", constraint: "> 1.1_alpha1", satisfied: true},
+		{version: "1.1", constraint: "< 1.1_alpha1", satisfied: false},
+		{version: "2.3.0b-r1", constraint: "< 2.3.0b-r2", satisfied: true},
+	}
+
+	for _, test := range tests {
+		t.Run(test.tName(), func(t *testing.T) {
+			constraint, err := GetConstraint(test.constraint, ApkFormat)
+
+			assert.NoError(t, err, "unexpected error from newApkConstraint: %v", err)
+			test.assertVersionConstraint(t, ApkFormat, constraint)
+
+		})
+	}
+}
+
+func TestApkVersion_Compare(t *testing.T) {
 	tests := []struct {
 		name           string
 		thisVersion    string
@@ -80,7 +133,7 @@ func TestApkVersionCompare(t *testing.T) {
 	}
 }
 
-func TestApkVersionCompareEdgeCases(t *testing.T) {
+func TestApkVersion_Compare_EdgeCases(t *testing.T) {
 	tests := []struct {
 		name           string
 		setupFunc      func(testing.TB) (*Version, *Version)
