+++ grype/version/error.go
@@ -43,3 +43,17 @@ func (e *UnsupportedComparisonError) Is(target error) bool {
 func invalidFormatError(format Format, raw string, err error) error {
 	return fmt.Errorf("invalid %s version from '%s': %w", format.String(), raw, err)
 }
+
+// NonFatalConstraintError should be used any time an unexpected but recoverable condition is encountered while
+// checking version constraint satisfaction. The error should get returned by any implementer of the Constraint
+// interface. If returned by the Satisfied method on the Constraint interface, this error will be caught and
+// logged as a warning in the FindMatchesByPackageDistro function in grype/matcher/common/distro_matchers.go
+type NonFatalConstraintError struct {
+	constraint Constraint
+	version    *Version
+	message    string
+}
+
+func (e NonFatalConstraintError) Error() string {
+	return fmt.Sprintf("matching raw constraint %s against version %s caused a non-fatal error: %s", e.constraint, e.version, e.message)
+}
