+++ grype/version/gem_version_test.go
@@ -9,6 +9,49 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestGemVersion_Constraint(t *testing.T) {
+	tests := []testCase{
+		// empty values
+		{version: "2.3.1", constraint: "", satisfied: true},
+		// typical cases
+		{version: "0.9.9-r0", constraint: "< 0.9.12-r1", satisfied: true}, // regression case
+		{version: "1.5.0-arm-windows", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: true},
+		{version: "0.2.0-arm-windows", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: true},
+		{version: "0.0.1-armv5-window", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.0.1-armv7-linux", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.6.0-universal-darwin-9", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.6.0-universal-darwin-10", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "0.6.0-x86_64-darwin-10", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "2.5.0", constraint: "> 0.1.0, < 0.5.0 || > 1.0.0, < 2.0.0", satisfied: false},
+		{version: "1.2.0", constraint: ">1.0, <2.0", satisfied: true},
+		{version: "1.2.0-x86", constraint: ">1.0, <2.0", satisfied: true},
+		{version: "1.2.0-x86-linux", constraint: ">1.0, <2.0", satisfied: true},
+		{version: "1.2.0-x86-linux", constraint: "= 1.2.0", satisfied: true},
+		{version: "1.2.0-x86_64-linux", constraint: "= 1.2.0", satisfied: true},
+		{version: "1.2.0-x86_64-linux", constraint: "< 1.2.1", satisfied: true},
+		// https://semver.org/#spec-item-11
+		{version: "1.2.0-alpha-x86-linux", constraint: "<1.2.0", satisfied: true},
+		{version: "1.2.0-alpha-1-x86-linux", constraint: "<1.2.0", satisfied: true},
+		// gem versions seem to respect the order: {sem-version}+{meta}-{arch}-{os}
+		// but let's check the extraction works even when the order of {meta}-{arch} varies.
+		{version: "1.2.0-alpha-1-x86-linux-meta", constraint: "<1.2.0", satisfied: true},
+		{version: "1.2.0-alpha-1-meta-x86-linux", constraint: "<1.2.0", satisfied: true},
+		{version: "1.2.0-alpha-1-x86-linux-meta", constraint: ">1.1.0", satisfied: true},
+		{version: "1.2.0-alpha-1-arm-linux-meta", constraint: ">1.1.0", satisfied: true},
+		{version: "1.0.0-alpha-a.b-c-somethinglong-build.1-aef.1-its-okay", constraint: "<1.0.0", satisfied: true},
+	}
+
+	for _, test := range tests {
+		t.Run(test.tName(), func(t *testing.T) {
+			constraint, err := GetConstraint(test.constraint, GemFormat)
+			assert.NoError(t, err, "unexpected error from newSemanticConstraint: %v", err)
+
+			test.assertVersionConstraint(t, GemFormat, constraint)
+		})
+	}
+
+}
+
 func Test_cleanPlatformMakesEqualVersions(t *testing.T) {
 	tests := []struct {
 		input   string
@@ -52,7 +95,7 @@ func Test_cleanPlatformMakesEqualVersions(t *testing.T) {
 	}
 }
 
-func TestNewRubyVersion_ValidInputs(t *testing.T) {
+func TestNewGemVersion_ValidInputs(t *testing.T) {
 	tests := []struct {
 		input              string
 		expectedOriginal   string // What v.original should be
@@ -84,7 +127,7 @@ func TestNewRubyVersion_ValidInputs(t *testing.T) {
 	}
 }
 
-func TestNewRubyVersion_InvalidInputs(t *testing.T) {
+func TestNewGemVersion_InvalidInputs(t *testing.T) {
 	invalidVersions := []struct {
 		name           string
 		input          string
@@ -115,7 +158,7 @@ func TestNewRubyVersion_InvalidInputs(t *testing.T) {
 	}
 }
 
-func TestRubyVersion_Compare(t *testing.T) {
+func TestGemVersion_Compare(t *testing.T) {
 	tests := []struct {
 		v1   string
 		v2   string
@@ -208,7 +251,7 @@ func TestRubyVersion_Compare(t *testing.T) {
 	}
 }
 
-func TestRubyVersion_Compare_Errors(t *testing.T) {
+func TestGemVersion_Compare_Errors(t *testing.T) {
 	vGem1_0, err := newGemVersion("1.0")
 	require.NoError(t, err)
 
@@ -240,7 +283,7 @@ func TestRubyVersion_Compare_Errors(t *testing.T) {
 	})
 }
 
-func TestRubyVersion_canonical(t *testing.T) {
+func TestGemVersion_canonical(t *testing.T) {
 	tests := []struct {
 		name    string
 		version string
