+++ grype/version/jvm_version_test.go
@@ -7,7 +7,70 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestVersionJVM(t *testing.T) {
+func TestJVMVersion_Constraint(t *testing.T) {
+	tests := []testCase{
+		// pre jep 223 versions
+		{version: "1.7.0_80", constraint: "< 1.8.0", satisfied: true},
+		{version: "1.8.0_131", constraint: "> 1.8.0", satisfied: true},
+		{version: "1.8.0_131", constraint: "< 1.8.0_132", satisfied: true},
+		{version: "1.8.0_131-b11", constraint: "< 1.8.0_132", satisfied: true},
+
+		{version: "1.7.0_80", constraint: "> 1.8.0", satisfied: false},
+		{version: "1.8.0_131", constraint: "< 1.8.0", satisfied: false},
+		{version: "1.8.0_131", constraint: "> 1.8.0_132", satisfied: false},
+		{version: "1.8.0_131-b11", constraint: "> 1.8.0_132", satisfied: false},
+
+		{version: "1.7.0_80", constraint: "= 1.8.0", satisfied: false},
+		{version: "1.8.0_131", constraint: "= 1.8.0", satisfied: false},
+		{version: "1.8.0_131", constraint: "= 1.8.0_132", satisfied: false},
+		{version: "1.8.0_131-b11", constraint: "= 1.8.0_132", satisfied: false},
+
+		{version: "1.8.0_80", constraint: "= 1.8.0_80", satisfied: true},
+		{version: "1.8.0_131", constraint: ">= 1.8.0_131", satisfied: true},
+		{version: "1.8.0_131", constraint: "= 1.8.0_131-b001", satisfied: true}, // builds should not matter
+		{version: "1.8.0_131-ea-b11", constraint: "= 1.8.0_131-ea", satisfied: true},
+
+		// jep 223 versions
+		{version: "8.0.4", constraint: "> 8.0.3", satisfied: true},
+		{version: "8.0.4", constraint: "< 8.0.5", satisfied: true},
+		{version: "9.0.0", constraint: "> 8.0.5", satisfied: true},
+		{version: "9.0.0", constraint: "< 9.1.0", satisfied: true},
+		{version: "11.0.4", constraint: "<= 11.0.4", satisfied: true},
+		{version: "11.0.5", constraint: "> 11.0.4", satisfied: true},
+
+		{version: "8.0.4", constraint: "< 8.0.3", satisfied: false},
+		{version: "8.0.4", constraint: "> 8.0.5", satisfied: false},
+		{version: "9.0.0", constraint: "< 8.0.5", satisfied: false},
+		{version: "9.0.0", constraint: "> 9.1.0", satisfied: false},
+		{version: "11.0.4", constraint: "> 11.0.4", satisfied: false},
+		{version: "11.0.5", constraint: "< 11.0.4", satisfied: false},
+
+		// mixed versions
+		{version: "1.8.0_131", constraint: "< 9.0.0", satisfied: true}, // 1.8.0_131 -> 8.0.131
+		{version: "9.0.0", constraint: "> 1.8.0_131", satisfied: true}, // 1.8.0_131 -> 8.0.131
+		{version: "1.8.0_131", constraint: "<= 8.0.131", satisfied: true},
+		{version: "1.8.0_131", constraint: "> 7.0.79", satisfied: true},
+		{version: "1.8.0_131", constraint: "= 8.0.131", satisfied: true},
+		{version: "1.8.0_131", constraint: ">= 9.0.0", satisfied: false},
+		{version: "9.0.1", constraint: "< 8.0.131", satisfied: false},
+
+		// pre-release versions
+		{version: "1.8.0_131-ea", constraint: "< 1.8.0_131", satisfied: true},
+		{version: "1.8.0_131", constraint: "> 1.8.0_131-ea", satisfied: true},
+		{version: "9.0.0-ea", constraint: "< 9.0.0", satisfied: true},
+		{version: "9.0.0-ea", constraint: "> 1.8.0_131", satisfied: true},
+	}
+
+	for _, test := range tests {
+		t.Run(test.version+"_constraint_"+test.constraint, func(t *testing.T) {
+			constraint, err := GetConstraint(test.constraint, JVMFormat)
+			require.NoError(t, err)
+			test.assertVersionConstraint(t, JVMFormat, constraint)
+		})
+	}
+}
+
+func TestJVMVersion_Compare(t *testing.T) {
 	tests := []struct {
 		v1       string
 		v2       string
@@ -61,21 +124,22 @@ func TestVersionJVM(t *testing.T) {
 	for _, test := range tests {
 		name := test.v1 + "_vs_" + test.v2
 		t.Run(name, func(t *testing.T) {
-			v1, err := newJvmVersion(test.v1)
+			v1, err := NewVersion(test.v1, JVMFormat)
 			require.NotNil(t, v1)
 			require.NoError(t, err)
 
-			v2, err := newJvmVersion(test.v2)
+			v2, err := NewVersion(test.v2, JVMFormat)
 			require.NotNil(t, v2)
 			require.NoError(t, err)
 
-			actual := v1.compare(v2)
+			actual, err := v1.Compare(v2)
+			require.NoError(t, err)
 			require.Equal(t, test.expected, actual)
 		})
 	}
 }
 
-func TestConvertNonCompliantSemver(t *testing.T) {
+func TestJVMVersion_ConvertNonCompliantSemver(t *testing.T) {
 	tests := []struct {
 		name     string
 		input    string
@@ -126,7 +190,7 @@ func TestConvertNonCompliantSemver(t *testing.T) {
 	}
 }
 
-func TestVersionJVM_invalid(t *testing.T) {
+func TestJVMVersion_Invalid(t *testing.T) {
 	tests := []struct {
 		name    string
 		version string
@@ -150,7 +214,7 @@ func TestVersionJVM_invalid(t *testing.T) {
 	}
 }
 
-func TestJvmVersionCompare_Formats(t *testing.T) {
+func TestJvmVersion_Compare_Formats(t *testing.T) {
 	tests := []struct {
 		name           string
 		thisVersion    string
@@ -220,7 +284,7 @@ func TestJvmVersionCompare_Formats(t *testing.T) {
 	}
 }
 
-func TestJvmVersionCompareEdgeCases(t *testing.T) {
+func TestJvmVersion_Compare_EdgeCases(t *testing.T) {
 	tests := []struct {
 		name           string
 		setupFunc      func(testing.TB) (*Version, *Version)
