+++ grype/version/range_expression.go
@@ -7,27 +7,27 @@ import (
 	"text/scanner"
 )
 
-type constraintExpression struct {
-	units [][]constraintUnit // only supports or'ing a group of and'ed groups
+type simpleRangeExpression struct {
+	units [][]rangeUnit // only supports or'ing a group of and'ed groups
 }
 
-func newConstraintExpression(phrase string) (constraintExpression, error) {
+func parseRangeExpression(phrase string) (simpleRangeExpression, error) {
 	orParts, err := scanExpression(phrase)
 	if err != nil {
-		return constraintExpression{}, fmt.Errorf("unable to create constraint expression from=%q : %w", phrase, err)
+		return simpleRangeExpression{}, fmt.Errorf("unable to create constraint expression from=%q : %w", phrase, err)
 	}
 
-	orUnits := make([][]constraintUnit, len(orParts))
+	orUnits := make([][]rangeUnit, len(orParts))
 	var fuzzyErr error
 	for orIdx, andParts := range orParts {
-		andUnits := make([]constraintUnit, len(andParts))
+		andUnits := make([]rangeUnit, len(andParts))
 		for andIdx, part := range andParts {
-			unit, err := parseUnit(part)
+			unit, err := parseRange(part)
 			if err != nil {
-				return constraintExpression{}, err
+				return simpleRangeExpression{}, err
 			}
 			if unit == nil {
-				return constraintExpression{}, fmt.Errorf("unable to parse unit: %q", part)
+				return simpleRangeExpression{}, fmt.Errorf("unable to parse unit: %q", part)
 			}
 			andUnits[andIdx] = *unit
 		}
@@ -35,19 +35,19 @@ func newConstraintExpression(phrase string) (constraintExpression, error) {
 		orUnits[orIdx] = andUnits
 	}
 
-	return constraintExpression{
+	return simpleRangeExpression{
 		units: orUnits,
 	}, fuzzyErr
 }
 
-func (c *constraintExpression) satisfied(format Format, version *Version) (bool, error) {
+func (c *simpleRangeExpression) satisfied(format Format, version *Version) (bool, error) {
 	oneSatisfied := false
 	for i, andOperand := range c.units {
 		allSatisfied := true
 		for j, andUnit := range andOperand {
 			result, err := version.Compare(&Version{
 				Format: format,
-				Raw:    andUnit.rawVersion,
+				Raw:    andUnit.version,
 			})
 			if err != nil {
 				return false, fmt.Errorf("uncomparable %T vs %q: %w", andUnit, version.String(), err)
